[SPTFC-11] moved the CurrentlyPlaying component to its own file and refactor a couple of imports on the Footer component
