+++ src/components/player/index.tsx
@@ -1,0 +1,4 @@
+import { AudioPlayer } from "./AudioPlayer";
+import { CurrentlyPlaying } from "./CurrentlyPlaying";
+
+export { AudioPlayer, CurrentlyPlaying };
