+++ src/components/reusable/index.tsx
@@ -5,6 +5,7 @@ import { SongItem } from "./SongItem";
 import { SongList } from "./SongList";
 import { SpeakerButton } from "./SpeakerButton";
 import { VerticalCard } from "./VerticalCard";
+import { MicIcon } from "./CustomIcons";
 
 export {
   BigPlayButton,
@@ -14,4 +15,5 @@ export {
   SongList,
   SpeakerButton,
   VerticalCard,
+  MicIcon,
 };
